+++ src/linter/rules/index.ts
@@ -9,6 +9,7 @@ import { InvalidModifiers } from './invalid-modifiers';
 import { InconsistentHintPlatforms } from './inconsistent-hint-platforms';
 import { SingleSelector } from './single-selector';
 import { UnknownPreProcessorDirectives } from './unknown-preprocessor-directives';
+import { NoShortRules } from './no-short-rules';
 
 export const defaultLinterRules = new Map<string, AnyLinterRule>([
     ['if-closed', IfClosed],
@@ -21,4 +22,5 @@ export const defaultLinterRules = new Map<string, AnyLinterRule>([
     ['invalid-domain-list', InvalidDomainList],
     ['invalid-modifiers', InvalidModifiers],
     ['inconsistent-hint-platforms', InconsistentHintPlatforms],
+    ['no-short-rules', NoShortRules],
 ]);